Extract reference box functionality into new class ReferenceBoxHelper
