+++ app/frontend/components/tool_panel/commentary_panel.js
@@ -20,7 +20,7 @@ const eventController = require('../../controllers/event_controller.js');
 const VerseBox = require("../../ui_models/verse_box.js");
 const { getPlatform } = require('../../helpers/ezra_helper.js');
 const VerseBoxHelper = require('../../helpers/verse_box_helper.js');
-const sectionLabelHelper = require('../../helpers/section_label_helper.js');
+const ReferenceBoxHelper = require('./reference_box_helper.js');
 
 /**
  * The CommentaryPanel component implements a tool panel that shows Bible commentaries for selected verses
@@ -61,6 +61,7 @@ class CommentaryPanel {
     });
 
     this._verseBoxHelper = new VerseBoxHelper();
+    this._referenceBoxHelper = new ReferenceBoxHelper(this.getMainContent(), this.getReferenceBox());
   }
 
   setRefreshBlocked(refreshBlocked) {
@@ -189,7 +190,7 @@ class CommentaryPanel {
     this.getBoxContent().innerHTML = commentaryContent;
     this.getMainContent().scrollTop = 0;
 
-    this.hideReferenceBox();
+    this._referenceBoxHelper.hideReferenceBox();
     this.getReferenceBox().innerHTML = "";
 
     this.applyParagraphs();
@@ -211,14 +212,14 @@ class CommentaryPanel {
     let referenceElements = this.getBoxContent().querySelectorAll('reference');
     referenceElements.forEach((reference) => {
       reference.addEventListener('click', (event) => {
-        this.handleReferenceClick(event);
+        this._referenceBoxHelper.handleReferenceClick(event);
       });
     });
 
     let scripRefElements = this.getBoxContent().querySelectorAll('.sword-scripref');
     scripRefElements.forEach((scripRef) => {
       scripRef.addEventListener('click', (event) => {
-        this.handleReferenceClick(event);
+        this._referenceBoxHelper.handleReferenceClick(event);
       });
     });
 
@@ -300,93 +301,6 @@ class CommentaryPanel {
         button.classList.add('fa-circle-chevron-right');
       });
     }
-  }
-
-  async handleReferenceClick(event) {
-    event.preventDefault();
-    event.stopPropagation();
-
-    let newTabOption = app_controller.optionsMenu._verseListNewTabOption;
-
-    await app_controller.verse_list_popup.initCurrentCommentaryXrefs(event.target);
-
-    if (app_controller.verse_list_popup.currentXrefs.length > 2 || platformHelper.isMobile()) {
-      this.hideReferenceBox();
-
-      if (newTabOption.isChecked) {
-        await app_controller.verse_list_popup.openVerseListInNewTab();
-      } else {
-        await app_controller.verse_list_popup.openVerseListPopup(event, 'COMMENTARY_DICT_XREFS');
-      }
-
-    } else if (app_controller.verse_list_popup.currentXrefs.length > 0) {
-      await this.renderReferenceVerses(app_controller.verse_list_popup.currentXrefs);
-      event.target.scrollIntoView({ block: "nearest" });
-    }
-  }
-
-  async renderReferenceVerses(references) {
-    const bibleTranslationId = app_controller.tab_controller.getTab().getBibleTranslationId();
-    let verses = await ipcNsi.getVersesFromReferences(bibleTranslationId, references);
-    let verseReferences = [];
-    let bibleBookStats = app_controller.text_controller.getBibleBookStatsFromVerses(verses);
-    let multipleBooks = Object.keys(bibleBookStats).length > 1;
-
-    let multipleVerses = verses.length > 1;
-    let verseContent = "";
-
-    verses.forEach((verse) => {
-      if (multipleVerses && !multipleBooks) {
-        verseContent += `<sup>${verse.verseNr}</sup>`;
-      }
-
-      verseContent += verse.content + "<br/>";
-
-      if (multipleBooks) {
-        let verseReference = verse.bibleBookShortTitle + ' ' + verse.chapter + window.reference_separator + verse.verseNr;
-        verseContent += verseReference + "<br/><br/>";
-      }
-
-      let shortVerseReference = verse.chapter + window.reference_separator + verse.verseNr;
-      verseReferences.push(shortVerseReference);
-    });
-
-    if (!multipleBooks) {
-      let currentBook = Object.keys(bibleBookStats)[0];
-      let formattedVerseList = await sectionLabelHelper.formatVerseList(verseReferences, currentBook, window.reference_separator);
-
-      verseContent += `<div>${currentBook} ${formattedVerseList}</div>`;
-    }
-
-    verseContent = '<div class="panel-content verse-text" style="margin-top: 1em;">' + verseContent + '</div>';
-
-    let tabIconTitle = i18n.t('bible-browser.open-new-tab');
-    let tabIcon = `<div class="tab-icon icon" title="${tabIconTitle}"><i class="fa-solid fa-arrow-up-right-from-square"></i></div>`;
-    let closeIcon = '<div class="close-icon icon"><i class="fa-solid fa-rectangle-xmark"></i></div>';
-
-    const commentaryPanelReferenceBox = this.getReferenceBox();
-    commentaryPanelReferenceBox.innerHTML = closeIcon + tabIcon + verseContent;
-
-    commentaryPanelReferenceBox.querySelector('.close-icon').addEventListener('click', (event) => {
-      this.hideReferenceBox();
-    });
-
-    commentaryPanelReferenceBox.querySelector('.tab-icon').addEventListener('click', (event) => {
-      this.hideReferenceBox();
-      app_controller.verse_list_popup.openVerseListInNewTab();
-    });
-
-    this.showReferenceBox();
-  }
-
-  showReferenceBox() {
-    this.getMainContent().classList.add('with-reference-box');
-    this.getReferenceBox().style.display = 'block';
-  }
-
-  hideReferenceBox() {
-    this.getMainContent().classList.remove('with-reference-box');
-    this.getReferenceBox().style.display = 'none';
   }
 
   processCommentaryHtml(htmlInput) {
